+++ runtime/t0rn-parachain/src/circuit_config.rs
@@ -346,7 +346,7 @@ parameter_types! {
     pub const EpochsPerSyncCommitteePeriod: u32 = 256;
     pub const HeadersToStoreEth: u32 = 50400 + 1; // 1 week + 1. We want a multiple of 32 + 1.
     pub const CommitteeMajorityThresholdEth2: u32 = 80;
-    pub const CommitteeMajorityThresholdSepolia: u32 = 1;
+    pub const CommitteeMajorityThresholdSepolia: u32 = 0;
 }
 
 impl pallet_eth2_finality_verifier::Config for Runtime {