+++ CHANGELOG.md
@@ -4,6 +4,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.65.0-rc.0] - 2023-10-22
+### :sparkles: New Features
+- [`b68ad72`](https://github.com/t3rn/t3rn/commit/b68ad72468b98ea299d159152387b3688013bbe3) - eth2-proof npm package *(PR [#1374](https://github.com/t3rn/t3rn/pull/1374) by [@3h4x](https://github.com/3h4x))*
+
+### :bug: Bug Fixes
+- [`cac2a09`](https://github.com/t3rn/t3rn/commit/cac2a09f9907f311b86ab8cc72dc722d6d61f233) - decrease CommitteeMajorityThresholdSepolia for sepolia ranger *(PR [#1379](https://github.com/t3rn/t3rn/pull/1379) by [@3h4x](https://github.com/3h4x))*
+
+
+## [v1.64.1-rc.0] - 2023-10-19
+### :bug: Bug Fixes
+- [`60fe33a`](https://github.com/t3rn/t3rn/commit/60fe33a391e3dbf36724bfe191d84cf40bc53e1c) - **executor**: error with transfer on rococo *(PR [#1376](https://github.com/t3rn/t3rn/pull/1376) by [@3h4x](https://github.com/3h4x))*
+- [`ad57e65`](https://github.com/t3rn/t3rn/commit/ad57e6571023d44d6f7ffb6e8d17ecc9a26ba442) - enable XCM transfers of USDT, ROC, and TRN between AssetHub & t0rn *(PR [#1375](https://github.com/t3rn/t3rn/pull/1375) by [@chexware](https://github.com/chexware))*
+
+
+## [v1.64.0-rc.0] - 2023-10-19
+### :sparkles: New Features
+- [`0075308`](https://github.com/t3rn/t3rn/commit/0075308b00b04606d9f3a88ae8ab87bb658feb8f) - **sdk**: added SDK support for XCM transfers and enabled TRN XCM transfers through CLI *(PR [#1365](https://github.com/t3rn/t3rn/pull/1365) by [@chexware](https://github.com/chexware))*
+- [`89f84ce`](https://github.com/t3rn/t3rn/commit/89f84ceb9902741a1356e70da39930f3b234ecd2) - add force flag for purge commands *(PR [#1370](https://github.com/t3rn/t3rn/pull/1370) by [@3h4x](https://github.com/3h4x))*
+
+### :bug: Bug Fixes
+- [`24e9c4a`](https://github.com/t3rn/t3rn/commit/24e9c4a6d9b4a5c0725b208352cf7ede4fdc293e) - runtime upgrade *(PR [#1367](https://github.com/t3rn/t3rn/pull/1367) by [@3h4x](https://github.com/3h4x))*
+- [`e5c07e9`](https://github.com/t3rn/t3rn/commit/e5c07e94a0c2735774b4d4db6c5d12dbd7198c8a) - remove matrix which messes up CI required checks *(PR [#1373](https://github.com/t3rn/t3rn/pull/1373) by [@3h4x](https://github.com/3h4x))*
+
+
 ## [v1.63.1-rc.0] - 2023-10-16
 ### :bug: Bug Fixes
 - [`5bc68dd`](https://github.com/t3rn/t3rn/commit/5bc68ddd12707e1b4d1b987c47b9340426c7ece3) - include binary for docker context *(PR [#1364](https://github.com/t3rn/t3rn/pull/1364) by [@3h4x](https://github.com/3h4x))*
@@ -1749,3 +1773,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v1.62.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.61.0-rc.0...v1.62.0-rc.0
 [v1.63.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.62.0-rc.0...v1.63.0-rc.0
 [v1.63.1-rc.0]: https://github.com/t3rn/t3rn/compare/v1.63.0-rc.0...v1.63.1-rc.0
+[v1.64.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.63.2-rc.0...v1.64.0-rc.0
+[v1.64.1-rc.0]: https://github.com/t3rn/t3rn/compare/v1.64.0-rc.0...v1.64.1-rc.0
+[v1.65.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.64.1-rc.0...v1.65.0-rc.0